@@ -3,29 +3,25 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "bootstrap": "^5.1.1",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.0.0-rc.0",
-    "react-chrono": "^1.12.1",
+    "bootstrap": "^5.3.2",
+    "matchmedia-polyfill": "^0.3.2",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.9.1",
+    "react-chrono": "^2.3.1",
     "react-dark-mode-toggle": "^0.2.0",
-    "react-dom": "^17.0.2",
-    "react-markdown": "^7.0.1",
-    "react-particles": "^2.12.2",
+    "react-dom": "^18.2.0",
+    "react-markdown": "^9.0.1",
     "react-reveal": "^1.2.2",
-    "react-router": "^5.2.1",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^6.20.0",
     "react-scripts": "^5.0.1",
-    "react-social-icons": "^5.6.1",
-    "styled-components": "^5.3.1",
-    "tsparticles-engine": "^2.12.0",
-    "typewriter-effect": "^2.18.2",
+    "react-social-icons": "^6.6.0",
+    "styled-components": "^6.1.1",
+    "typescript": "^5.3.2",
+    "typewriter-effect": "^2.21.0",
     "use-dark-mode": "^2.3.1",
     "vertical-timeline-component-for-react": "^1.0.7",
-    "web-vitals": "^1.0.1"
+    "web-vitals": "^3.5.0"
   },
   "browserslist": {
     "production": [
@@ -46,13 +42,13 @@
     "eject": "NODE_OPTIONS=--openssl-legacy-provider react-scripts eject"
   },
   "devDependencies": {
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "react-error-overlay": "6.0.9"
+    "eslint": "^8.54.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "react-error-overlay": "6.0.11"
   },
   "resolutions": {
     "react-error-overlay": "6.0.9"
