@@ -1,5 +1,5 @@
 import React, { useState, useEffect, Suspense } from 'react';
-import { Switch, Route } from 'react-router-dom';
+import { Routes, Route } from 'react-router-dom';
 import FallbackSpinner from './components/FallbackSpinner';
 import NavBarWithRouter from './components/NavBar';
 import Home from './components/Home';
@@ -21,9 +21,9 @@ function MainApp() {
     <div className="MainApp">
       <NavBarWithRouter />
       <main className="main">
-        <Switch>
-          <Suspense fallback={<FallbackSpinner />}>
-            <Route exact path="/" component={Home} />
+        <Suspense fallback={<FallbackSpinner />}>
+          <Routes>
+            <Route exact="true" path="/" element={<Home />} />
             {data
               && data.sections.map((route) => {
                 const SectionComponent = React.lazy(() => import('./components/' + route.component));
@@ -31,14 +31,12 @@ function MainApp() {
                   <Route
                     key={route.headerTitle}
                     path={route.path}
-                    component={() => (
-                      <SectionComponent header={route.headerTitle} />
-                    )}
+                    element={<SectionComponent header={route.headerTitle} />}
                   />
                 );
               })}
-          </Suspense>
-        </Switch>
+          </Routes>
+        </Suspense>
       </main>
     </div>
   );
