@@ -1,14 +1,16 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(
-  <React.StrictMode>
+ReactDOM.createRoot(
+  document.getElementById("root"),
+)
+.render(
+  <React.Fragment>
     <App />
-  </React.StrictMode>,
-  document.getElementById('root'),
+  </React.Fragment>,
 );
 
 // If you want to start measuring performance in your app, pass a function
