@@ -1,6 +1,5 @@
 import React, { useEffect, useState, useContext } from 'react';
 import { SocialIcon } from 'react-social-icons';
-import { ThemeContext } from 'styled-components';
 import endpoints from '../constants/endpoints';
 
 const styles = {
@@ -12,7 +11,6 @@ const styles = {
 };
 
 function Social() {
-  const theme = useContext(ThemeContext);
   const [data, setData] = useState(null);
 
   useEffect(() => {
@@ -30,9 +28,8 @@ function Social() {
         <SocialIcon
           key={social.network}
           style={styles.iconStyle}
-          url={social.href}
+          href={social.href}
           network={social.network}
-          bgColor={theme.socialIconBgColor}
           target="_blank"
           rel="noopener"
         />
